Pembenahan kode dan struktur kode

+++ WindowsFormsApp1/FormApoteker.cs
@@ -35,6 +35,12 @@ namespace ManajemenObatApp
 
         private void btnTambah_Click(object sender, EventArgs e)
         {
+            if (txtId.Text == "" || txtNama.Text == "" || txtPassword.Text.Length < 8)
+            {
+                MessageBox.Show("Isi data dengan benar!", "Peringatan", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                return;
+            }
+
             string query = "INSERT INTO apoteker VALUES (@id, @nama, @pass)";
             SqlParameter[] param = {
                 new SqlParameter("@id", txtId.Text),
@@ -53,6 +59,12 @@ namespace ManajemenObatApp
 
         private void btnHapus_Click(object sender, EventArgs e)
         {
+            if (txtId.Text == "")
+            {
+                MessageBox.Show("Ketik ID yang ingin dihapus!", "Peringatan", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                return;
+            }
+
             string query = "DELETE FROM apoteker WHERE id_apoteker=@id";
             SqlParameter[] param = {
                 new SqlParameter("@id", txtId.Text)
@@ -64,6 +76,12 @@ namespace ManajemenObatApp
 
         private void btnEdit_Click(object sender, EventArgs e)
         {
+            if(txtId.Text == "" || txtNama.Text == "" || txtPassword.Text.Length < 8)
+            {
+                MessageBox.Show("Isi data dengan benar!", "Peringatan", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                return;
+            }
+
             string query = "UPDATE apoteker SET nama=@nama, password=@pass WHERE id_apoteker=@id";
             SqlParameter[] param = {
                 new SqlParameter("@id", txtId.Text),
